Move IdentityLearner to its own file
